fixed stupid periodic checkpointing bug i introduced with COD.  :( the problem came from checking an int return value like a bool, when it was using UNIX syscall semantics (-1 for failure, 0 for success). now, all the kill-related methods return a C++ bool, so there's no confusion about return value semantics...

+++ src/condor_startd.V6/Resource.C
@@ -149,7 +149,7 @@ Resource::kill_claim( void )
 		return change_state( owner_state );
 	default:
 			// In other states, try direct kill.  See above.
-		return r_cur->starterKillHard();
+		return (int)r_cur->starterKillHard();
 	}
 	return TRUE;
 }
@@ -196,7 +196,7 @@ int
 Resource::request_new_proc( void )
 {
 	if( state() == claimed_state && r_cur->isActive()) {
-		return r_cur->starterKill( SIGHUP );
+		return (int)r_cur->starterKill( SIGHUP );
 	} else {
 		return FALSE;
 	}
